* Attempts to automatically reload instant answers upon detect changes in the file system.
Initial commit is WIP and broken due to not being able to run App::DuckPAN::DDG::get_blocks_from_current_dir() multiple times in the same process.  This is what actually createes the list of IAs to be test. So the forking method ultimately may not work with the existing App::DuckPAN